% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_naive_bayes.R
\name{predict.tidy_naive_bayes}
\alias{predict.tidy_naive_bayes}
\title{Predict method for Naive Bayes classifier models}
\usage{
\method{predict}{tidy_naive_bayes}(model, newdata = NULL, type = c("class",
  "prob"))
}
\description{
Predict method for Naive Bayes classifier models
}
