% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_naive_bayes.R
\name{tidy_naive_bayes}
\alias{tidy_naive_bayes}
\alias{tidy_naive_bayes.default}
\alias{tidy_naive_bayes.formula}
\alias{tidy_naive_bayes.tbl_df}
\title{Tidy Naive Bayes Classifier}
\usage{
\method{tidy_naive_bayes}{default}(x, y)

\method{tidy_naive_bayes}{formula}(formula, data)

\method{tidy_naive_bayes}{tbl_df}(.data, class_var)
}
\arguments{
\item{x}{A data frame.}

\item{y}{A vector.}

\item{formula}{An object of class 'formula' of the form 'class ~ predictors'.}

\item{data}{A data frame.}
}
\value{
A Tidy Naive Bayes classifier model.
}
\description{
Tidy Naive Bayes Classifier
}
\examples{
library(tibble)
data(iris)
model <- tidy_naive_bayes(Species ~ ., iris)

new_flowers <- tribble(
~Sepal.Length, ~Sepal.Width, ~Petal.Length, ~Petal.Width,
          5.3,          3.4,           1.6,          0.3,
          7.7,          3.9,           6.3,          2.1
)

predict(model, new_flowers)
predict(model, new_flowers, type = "prob")
}
\seealso{
\code{\link{predict.tidy_naive_bayes}}
}
