% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_naive_bayes.R
\name{predict.tidy_naive_bayes}
\alias{predict.tidy_naive_bayes}
\title{Predict class of new data using a Tidy Naive Bayes classifier model}
\usage{
\method{predict}{tidy_naive_bayes}(model, newdata = NULL, type = c("class",
  "prob"))
}
\description{
Predict class of new data using a Tidy Naive Bayes classifier model
}
