% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_naive_bayes.R
\name{tidy_naive_bayes}
\alias{tidy_naive_bayes}
\alias{tidy_naive_bayes.default}
\alias{tidy_naive_bayes.formula}
\alias{tidy_naive_bayes.tbl_df}
\title{Tidy Naive Bayes Classifier}
\usage{
\method{tidy_naive_bayes}{default}(x, y)

\method{tidy_naive_bayes}{formula}(formula, data)

\method{tidy_naive_bayes}{tbl_df}(.data, class_var)
}
\arguments{
\item{x}{A data frame.}

\item{y}{A vector.}

\item{formula}{An object of class 'formula' of the form 'class ~ predictors'.}

\item{data}{A data frame.}
}
\description{
Tidy Naive Bayes Classifier
}
\examples{
data(iris)
model <- tidy_naive_bayes(Species ~ ., iris)
}
\seealso{
[predict.tidy_naive_bayes()]
}
